Add files via upload
Project description
This project “Bank Management System in c Programming” is complete and totally-error free clean code. The main goal of this project is to create, alter or delete the customer account within this system. It is also named as “Customer Account Bank Management System”.
In this Management System, the user can add the new account, update that account, view and manage transaction within that particular account, checks whether the customer details are saved or not, can be used to view all the customer name list and also remove the customer’s account if it’s no longer needed.
The source code of this “Bank management System” is about 600 lines. Also, it is password protected. Hence, this project is just like real banking software with all basic features. It doesn’t consist of high graphics, It is the simple one and users can understand very easily.
The source code for Customer Account Bank Management System is fairly brief and easy to understand. we have divided this project into lots of functions, the majority of which belong to different banking activities. Listed below are some of the more vital functions which might help you understand this project much better.
menu() – The menu() function is used to display the menu where a user can apply different banking activities which are further described below.

new_acc() – The new_acc() function is used to create a new customer account in the system, where the customer details are saves such as name, dob, citizenship number, address and contact number. also, a customer can choose their account type (current account or fixed account).

edit() – The edit() function is used to customize or alter the details of a customer in that system.

transact() – The transact() function is used to deposit or withdraw money to and from a particular customer account.

view_list() – The view_list() function is used to view the customer details such as name, address, contact number etc.

erase() – The erase() function is used to delete the customer information or account from that system.

see() – The see() function is just like a view_list function but the different part is here a user can see the customer details with their account transactions.